requirements.txt
# Core dependencies
python-dotenv>=1.0
pyyaml>=6.0
pillow>=10.4

# LLM clients
google-generativeai>=0.8.3
huggingface_hub[inference]>=0.24

# Dataset loading
datasets>=2.14
